Adds start input event handlers for Game Engine
The game engine interface has GameEngine.StartInputHandler
which can be used to specify the conditions under which the skill should
be notified of Button events.
The GameEngine.InputHandlerEvent recceives Button events from the Alexa
service.

+++ src/com/amazon/speech/speechlet/interfaces/gameengine/Parameters.java
@@ -5,6 +5,6 @@ import java.util.List;
 public class Parameters {
 
     public String triggerEvent;
-    public Double triggerEventTimeMs;
+    public int triggerEventTimeMs;
     public List<Animation> animations;
 }
