+++ src/com/amazon/speech/speechlet/interfaces/gameengine/Sequence.java
@@ -2,7 +2,7 @@ package com.amazon.speech.speechlet.interfaces.gameengine;
 
 public class Sequence {
 
-    public Double durationMs;
+    public int durationMs;
     public boolean blend;
     public String color;
     public int intensity;
